split gui module
