+++ modules/allratings-gui-bukkit/build.gradle.kts
@@ -1,0 +1,21 @@
+plugins {
+    kotlin("jvm")
+    kotlin("plugin.serialization")
+    alias(libs.plugins.klibs.gradle.java.core)
+}
+dependencies {
+    // Kotlin
+    implementation(libs.bundles.kotlin)
+    // AstraLibs
+    implementation(libs.minecraft.astralibs.core)
+    // klibs
+    implementation(libs.klibs.kdi)
+    implementation(libs.klibs.mikro.core)
+    // Test
+    testImplementation(libs.bundles.testing.kotlin)
+    testImplementation(libs.tests.kotlin.test)
+    // Local
+    implementation(projects.modules.apiRating)
+    implementation(projects.modules.dbRating)
+    implementation(projects.modules.core)
+}
