+++ modules/api-rating/src/main/kotlin/ru/astrainteractive/astrarating/api/rating/di/ApiRatingModule.kt
@@ -8,7 +8,6 @@ import ru.astrainteractive.astrarating.api.rating.api.impl.CachedApiImpl
 import ru.astrainteractive.astrarating.api.rating.api.impl.RatingDBApiImpl
 import ru.astrainteractive.klibs.kdi.Provider
 import ru.astrainteractive.klibs.kdi.getValue
-import java.io.File
 
 interface ApiRatingModule {
     val ratingDBApi: RatingDBApi
@@ -17,13 +16,11 @@ interface ApiRatingModule {
     class Default(
         database: Database,
         coroutineScope: CoroutineScope,
-        pluginFolder: File
     ) : ApiRatingModule {
 
         override val ratingDBApi: RatingDBApi by Provider {
             RatingDBApiImpl(
                 database = database,
-                pluginFolder = pluginFolder
             )
         }
         override val cachedApi: CachedApi by Provider {
