+++ modules/core/src/main/kotlin/ru/astrainteractive/astrarating/core/RatingPermission.kt
@@ -7,7 +7,5 @@ sealed class RatingPermission(override val value: String) : Permission {
     data object MaxRatePerDay : RatingPermission("astra_rating.max_rate_per_day")
     data object SinglePlayerPerDay : RatingPermission("astra_rating.single_player_rate_per_day")
     data object Vote : RatingPermission("astra_rating.vote")
-    data object SeeRecords : RatingPermission("astra_rating.see_records")
-    data object MergeRecords : RatingPermission("astra_rating.merge_records")
     data object DeleteReport : RatingPermission("delete_report.vote")
 }
