+++ modules/gui-core-bukkit/src/main/kotlin/ru/astrainteractive/astrarating/gui/slot/SlotContext.kt
@@ -1,0 +1,9 @@
+package ru.astrainteractive.astrarating.gui.slot
+
+import ru.astrainteractive.astrarating.core.EmpireConfig
+import ru.astrainteractive.astrarating.core.PluginTranslation
+
+internal interface SlotContext {
+    val translation: PluginTranslation
+    val config: EmpireConfig
+}
