+++ modules/gui-core-bukkit/src/main/kotlin/ru/astrainteractive/astrarating/gui/util/TranslationDesc.kt
@@ -1,0 +1,8 @@
+package ru.astrainteractive.astrarating.gui.util
+
+import ru.astrainteractive.astralibs.string.StringDesc
+import ru.astrainteractive.astrarating.core.PluginTranslation
+
+internal class TranslationDesc(private val provider: (PluginTranslation) -> StringDesc.Raw) {
+    fun toString(pluginTranslation: PluginTranslation) = provider.invoke(pluginTranslation)
+}
