+++ settings.gradle.kts
@@ -48,3 +48,4 @@ include("modules:api-rating")
 include("modules:integration-papi")
 include("modules:shared")
 include("modules:core")
+include("modules:gui-core-bukkit")
