+++ spigot/build.gradle.kts
@@ -40,6 +40,7 @@ dependencies {
     implementation(projects.modules.integrationPapi)
     implementation(projects.modules.shared)
     implementation(projects.modules.core)
+    implementation(projects.modules.guiCoreBukkit)
 }
 minecraftMultiplatform {
     dependencies {
