--- spigot/src/main/kotlin/ru/astrainteractive/astrarating/gui/playerratings/di/PlayerRatingGuiDependencies.kt
@@ -1,15 +1,0 @@
-package ru.astrainteractive.astrarating.gui.playerratings.di
-
-import ru.astrainteractive.astralibs.kyori.KyoriComponentSerializer
-import ru.astrainteractive.astrarating.api.rating.api.RatingDBApi
-import ru.astrainteractive.astrarating.core.EmpireConfig
-import ru.astrainteractive.astrarating.core.PluginTranslation
-import ru.astrainteractive.klibs.mikro.core.dispatchers.KotlinDispatchers
-
-interface PlayerRatingGuiDependencies {
-    val dbApi: RatingDBApi
-    val dispatchers: KotlinDispatchers
-    val translation: PluginTranslation
-    val config: EmpireConfig
-    val translationContext: KyoriComponentSerializer
-}
